api/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
aiohttp
pydantic
